creating an dependency injection layer and removing assembler responsability from ManagerViewController
