+++ bubble-jam/Source/Main/Protocol/Assembler.swift
@@ -1,0 +1,13 @@
+//
+//  Assembler.swift
+//  bubble-jam
+//
+//  Created by Thiago Henrique on 25/07/23.
+//
+
+import Foundation
+
+protocol Assembler {
+    associatedtype ViewController
+    static func make() -> ViewController
+}
